+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.27.0] - 2025-11-11
+
 ### Added
 
 - Added missing resource_dfn_clone() arguments to Resource.clone()